requirements.txt
flask==3.0.2
diffusers==0.30.3
transformers==4.45.2
accelerate==0.30.1
safetensors==0.4.3
pillow      